add Navigation Component in ConnectWalletActivity because of WalletConnectModal integration
