+++ example/src/main/java/org/xmtp/android/example/connect/ConnectWalletActivity.kt
@@ -20,11 +20,6 @@ import org.xmtp.android.example.databinding.ActivityConnectWalletBinding
 
 class ConnectWalletActivity : AppCompatActivity() {
 
-    companion object {
-        private const val WC_URI_SCHEME = "wc://wc?uri="
-    }
-
-    private val viewModel: ConnectWalletViewModel by viewModels()
     private lateinit var binding: ActivityConnectWalletBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -33,72 +28,10 @@ class ConnectWalletActivity : AppCompatActivity() {
         binding = ActivityConnectWalletBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.uiState.collect(::ensureUiState)
-            }
-        }
 
-        binding.generateButton.setOnClickListener {
-            viewModel.generateWallet()
-        }
 
-        val isConnectWalletAvailable = isConnectAvailable()
-        binding.connectButton.isEnabled = isConnectWalletAvailable
-        binding.connectError.isVisible = !isConnectWalletAvailable
-        binding.connectButton.setOnClickListener {
-            binding.connectButton.start(viewModel.walletConnectKit, ::onConnected, ::onDisconnected)
-        }
+
     }
 
-    private fun onConnected(address: String) {
-        viewModel.connectWallet()
-    }
 
-    private fun onDisconnected() {
-        // No-op currently.
-    }
-
-    private fun isConnectAvailable(): Boolean {
-        val wcIntent = Intent(Intent.ACTION_VIEW).apply {
-            data = Uri.parse(WC_URI_SCHEME)
-        }
-        return wcIntent.resolveActivity(packageManager) != null
-    }
-
-    private fun ensureUiState(uiState: ConnectWalletViewModel.ConnectUiState) {
-        when (uiState) {
-            is ConnectWalletViewModel.ConnectUiState.Error -> showError(uiState.message)
-            ConnectWalletViewModel.ConnectUiState.Loading -> showLoading()
-            is ConnectWalletViewModel.ConnectUiState.Success -> signIn(
-                uiState.address,
-                uiState.encodedKeyData
-            )
-            ConnectWalletViewModel.ConnectUiState.Unknown -> Unit
-        }
-    }
-
-    private fun signIn(address: String, encodedKey: String) {
-        val accountManager = AccountManager.get(this)
-        Account(address, resources.getString(R.string.account_type)).also { account ->
-            accountManager.addAccountExplicitly(account, encodedKey, null)
-        }
-        startActivity(Intent(this, MainActivity::class.java))
-        finish()
-    }
-
-    private fun showError(message: String) {
-        binding.progress.visibility = View.GONE
-        binding.generateButton.visibility = View.VISIBLE
-        binding.connectButton.visibility = View.VISIBLE
-        binding.connectError.isVisible = !isConnectAvailable()
-        Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
-    }
-
-    private fun showLoading() {
-        binding.progress.visibility = View.VISIBLE
-        binding.generateButton.visibility = View.GONE
-        binding.connectButton.visibility = View.GONE
-        binding.connectError.visibility = View.GONE
-    }
 }
